requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
gunicorn
whitenoise
dj-database-url
psycopg2-binary
razorpay
